% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajustar_hora.R
\name{ajustar_hora}
\alias{ajustar_hora}
\title{Corrige el tiempo de acuerdo a la zona horaria.}
\usage{
ajustar_hora(
  GPS_raw = GPS_raw,
  dif_hor = dif_hor,
  col_dia = col_dia,
  col_hora = col_hora,
  formato = formato
)
}
\arguments{
\item{GPS_raw}{tu data frame con columna de hora y tiempo}

\item{dif_hor}{un numero correspondiente a la diferencia horaria}

\item{col_dia}{el nombre de la columna donde tienes la fecha}

\item{col_hora}{el nombre de la columna donde tienes el tiempo}

\item{formato}{el formato en el que esta tu fecha y hora}
}
\value{
un data frame con dos columnas adicionales dia_hora y hora_corregida
}
\description{
Corrige el tiempo de acuerdo a la zona horaria.
}
\examples{
t_formato="\%d/\%m/\%Y \%H:\%M:\%S"
GPS_gmt<-ajustar_hora(GPS_raw = GPS_raw, 
dif_hor = 5,
col_dia = 'DateGMT',
col_hora = 'TimeGMT',
formato=t_formato)
}
