% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_duracion.R
\name{calcular_duracion}
\alias{calcular_duracion}
\title{Calcula la duración de los viajes de alimentación}
\usage{
calcular_duracion(
  GPS_data = GPS_data,
  col_diahora = col_diahora,
  formato = formato,
  unidades = unidades,
  separador = separador
)
}
\arguments{
\item{GPS_data}{un data frame con una columna que incluya día y hora, y otra columna que incluya el numero del viaje}

\item{col_diahora}{la columna en formato POSTIXct con información del día y hora}

\item{formato}{el formato en el que esta la hora y día en la columna POSTIXct con información del día y hora}

\item{unidades}{elegir "hours", "minutes", "seconds".}

\item{separador}{la columna a usar para separar los viajes, puede ser el numero del viaje o separar por individuos, por ejemplo 'trip_number'}
}
\value{
regresa un nuevo data frame con la información del viaje, cuando inicio, cuando termino y la duración
}
\description{
Calcula la duración de los viajes de alimentación
}
\examples{
formato<-"\%Y-\%m-\%d \%H:\%M:\%S"
duracion<-calcular_duracion(GPS_data=GPS_edited,col_diahora="tStamp",
formato=formato,unidades="hours",separador='trip_number')
}
