% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_tripparams.R
\name{calcular_tripparams}
\alias{calcular_tripparams}
\title{Calcula los parametros del viaje de alimentacion}
\usage{
calcular_tripparams(
  GPS_edited = GPS_edited,
  col_diahora = col_diahora,
  formato = formato,
  nest_loc = nest_loc
)
}
\arguments{
\item{GPS_edited}{es un data frame con tus datos, debe contener columnas 'Longitude','Latitude' y una columna en formato POSTIXct}

\item{col_diahora}{una columna en formato POSTIXct}

\item{formato}{el formato en el que esta la col_diahora}

\item{nest_loc}{un data frame con las coordenadas del nido 'Longitude','Latitude'}
}
\value{
una tabla con los parametros del viaje
}
\description{
Calcula los parametros del viaje de alimentacion
}
\examples{
nest_loc=data.frame(Longitude = -109.4531,Latitude = -27.20097)
trip_params<-calcular_tripparams(GPS_edited = GPS_edited,
col_diahora = "tStamp",
formato = "\%Y-\%m-\%d \%H:\%M:\%S",
nest_loc=nest_loc)
}
