% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contar_viajes.R
\name{contar_viajes}
\alias{contar_viajes}
\title{Cuenta viajes y elimina periodos que no sean viajes de alimentacion}
\usage{
contar_viajes(GPS_data = GPS_data)
}
\arguments{
\item{GPS_data}{un data frame con una columna llamada trip que incluya Y o N}
}
\value{
un data frame con una columna con trip_number, num_seq y solo las locaciones fuera de la colonia
}
\description{
Cuenta viajes y elimina periodos que no sean viajes de alimentacion
}
\examples{
GPS_viaje<-contar_viajes(GPS_data=GPS_trip)
}
