% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_trips.R
\name{count_trips}
\alias{count_trips}
\title{Adds a column with the trip number and removes central locations}
\usage{
count_trips(GPS_data = GPS_data)
}
\arguments{
\item{GPS_data}{A data frame with a column named 'trip_number'}
}
\value{
Returns a data frame with the trip number and without central locations
}
\description{
Adds a column with the trip number and removes central locations
}
\examples{
GPS_trips<-count_trips(GPS_data=GPS_trip)
}
