% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_puntos.R
\name{dist_puntos}
\alias{dist_puntos}
\title{Agrega una columna con la distancia entre cada punto.}
\usage{
dist_puntos(GPS_data = GPS_data)
}
\arguments{
\item{GPS_data}{es un data frame con datos de Longitude y Latitude}
}
\value{
el mismo data frame con una columna llamada pointsdist_km
}
\description{
Agrega una columna con la distancia entre cada punto.
}
\examples{
GPS_dist<-dist_puntos(GPS_data = GPS_edited)
}
