% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS01.R
\docType{data}
\name{GPS01}
\alias{GPS01}
\title{Tracks de bobo/piquero/kena Sula dactylatra en Rapa Nui}
\format{
Data frame contiene datos de un individuo.
Las columnas son Latitude, Longitude, DateGMT, TimeGMT, tStamp e ID.
Las horas ya estan corregidas.
}
\usage{
data(GPS01)
}
\description{
Datos de tracks de Sula dactylatra colectados usando GPS en Rapa Nui
Incluye 1038 observaciones de 6 variables
Incluye datos de 1 individuo
}
\examples{
data(GPS01)
}
\references{
Lerma et al. (2020)
}
\keyword{datasets}
