% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS01_trips.R
\docType{data}
\name{GPS01_trips}
\alias{GPS01_trips}
\title{Track from one booby}
\format{
Columns are Latitude, Longitude, DateGMT, TimeGMT, Ids, trip, trip_number
}
\usage{
data(GPS01_trips)
}
\description{
Tracking data from one Masked booby tracked at Rapa Nui
Includes 1038 obs and 6 variables
}
\examples{
data(GPS01_trips)
}
\references{
Lerma et al. (2020)
}
\keyword{datasets}
