% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS_edited.R
\docType{data}
\name{GPS_edited}
\alias{GPS_edited}
\title{Tracks de bobo/piquero/kena Sula dactylatra en Rapa Nui}
\format{
Un data frame en formato rda, contiene datos de cinco individuos.
Las columnas son Latitude, Longitude, DateGMT, TimeGMT, ID, tStampt, trip, num_sew, trip_number.
}
\usage{
data(GPS_edited)
}
\description{
Datos de tracks de Sula dactylatra colectados usando GPS en Rapa Nui
Incluye 886 observaciones de 9 variables
Incluye datos de 1 individuo
}
\examples{
data(GPS_edited)
}
\references{
Lerma et al. (2020)
}
\keyword{datasets}
