% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS_preparado.R
\docType{data}
\name{GPS_preparado}
\alias{GPS_preparado}
\title{Tracks de bobo/piquero/kena Sula dactylatra en Rapa Nui}
\format{
Un data frame en formato rda, contiene datos de diez individuos.
Las columnas son IDs, trip_number, num_seq, Latitude, Longitude, DateGMT, TimeGMT, maxdist_km, pointsdist_km, dia_hora, hora_corregida.
}
\usage{
data(GPS_preparado)
}
\description{
Datos de tracks de Sula dactylatra colectados usando GPS en Rapa Nui
Incluye 1980 observaciones de 11 variables
Incluye datos de 10 individuos
}
\examples{
data(GPS_preparado)
}
\references{
Lerma et al. (2020)
}
\keyword{datasets}
