% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS_raw.R
\docType{data}
\name{GPS_raw}
\alias{GPS_raw}
\title{Tracks de bobo/piquero/kena Sula dactylatra en Rapa Nui}
\format{
Un csv en formato rda, contiene datos de cinco individuos. Las columnas son Latitude, Longitude, DateGMT, TimeGMT e ID. Tratando de replicar tal cual se descargan de los dispositivos. Las horas no estan corregidas.
}
\usage{
data(GPS_raw)
}
\description{
Datos de tracks de Sula dactylatra colectados usando GPS en Rapa Nui
Incluye 10440 observaciones de 5 variables
Incluye datos de 5 individuos
}
\examples{
data(GPS_raw)
}
\references{
Lerma et al. (2020)
}
\keyword{datasets}
