% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS_trip.R
\docType{data}
\name{GPS_trip}
\alias{GPS_trip}
\title{Tracks de bobo/piquero/kena Sula dactylatra en Rapa Nui}
\format{
Un csv en formato rda, contiene datos de cinco individuos.
Las columnas son Latitude, Longitude, DateGMT, TimeGMT, ID, tStampt y trip.
}
\usage{
data(GPS_trip)
}
\description{
Datos de tracks de Sula dactylatra colectados usando GPS en Rapa Nui
Incluye 886 observaciones de 9 variables
Incluye datos de 1 individuo
}
\examples{
data(GPS_trip)
}
\references{
Lerma et al. (2020)
}
\keyword{datasets}
