% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPSdata.R, R/suladata.R
\docType{data}
\name{GPSdata}
\alias{GPSdata}
\title{Tracks de bobo/piquero/kena Sula dactylatra en Rapa Nui}
\format{
Un csv en formaro rda

Un csv en formaro rda.
}
\usage{
data(GPSdata)

data(GPSdata)
}
\description{
Datos de tracks de Sula dactylatra colectados usando GPS en Rapa Nui
Incluye 10440 observaciones de 5 variables
Incluye datos de 5 individuos

Datos de tracks de Sula dactylatra colectados usando GPS en Rapa Nui
Incluye 10440 observaciones de 5 variables
Incluye datos de 5 individuos
}
\examples{
data(GPSdata)
data(GPSdata)
}
\references{
Lerma et al. (2020)

Lerma et al. (2020)
}
\keyword{datasets}
