% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPSdata2.R
\docType{data}
\name{GPSdata2}
\alias{GPSdata2}
\title{Tracks de bobo/piquero/kena Sula dactylatra en Rapa Nui}
\format{
Un csv en formato rda, contiene datos de cinco individuos.
Las columnas son Latitude, Longitude, DiaHora e ID.
En este data frame la columna DiaHora esta en formato POSIXct
}
\usage{
data(GPSdata2)
}
\description{
Datos de tracks de Sula dactylatra colectados usando GPS en Rapa Nui
Incluye 10440 observaciones de 5 variables
Incluye datos de 5 individuos
}
\examples{
data(GPSdata2)
}
\references{
Lerma et al. (2020)
}
\keyword{datasets}
