% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identificar_viajes.R
\name{identificar_viajes}
\alias{identificar_viajes}
\title{Agrega columna identificando viajes}
\usage{
identificar_viajes(GPS_data = GPS_data, nest_loc = nest_loc, distancia_km = 1)
}
\arguments{
\item{GPS_data}{es el data frame con datos de GPS con las columnas Latitude y Longitude}

\item{nest_loc}{es un data frame con las columnas Latitude y Longitude}

\item{distancia_km}{un numero de acuerdo al buffer de interes}
}
\value{
el data frame con una columna adicional
}
\description{
Agrega columna identificando viajes
}
\examples{
nest_loc<-data.frame(Longitude=-109.4531, Latitude=-27.20097)
GPS_01_trips<-identificar_viajes(GPS_data=GPS_01,nest_loc=nest_loc,distancia_km=0.01)
}
