% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_trips.R
\name{identify_trips}
\alias{identify_trips}
\title{Adds a column with information when the animal was close to the central location or outside the central location}
\usage{
identify_trips(
  GPS_data = GPS_data,
  Nest_location = Nest_location,
  Distance_km = Distance_km
)
}
\arguments{
\item{GPS_data}{is the data frame with the tracking data, must have columns named 'Longitude' and 'Latitude' if not provided please rename those columns}

\item{Nest_location}{is the data frame with the nest location, must have columns named 'Longitude' and 'Latitude' if not provided please rename those columns}

\item{Distance_km}{is the distance that would be considered inside the central location or outside the central location}
}
\value{
returns the same data frame with an additional column, Y for outside the central location, N for inside the central location
}
\description{
Adds a column with information when the animal was close to the central location or outside the central location
}
\examples{
GPS_01_nest<-data.frame(Longitude=-109.4531, Latitude=-27.20097)
GPS01_trips<-identify_trips(GPS_data=GPS01,Nest_location=GPS_01_nest,Distance_km=1)
}
