% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{A function to interpolate your data at an specific interval}
\usage{
interpolate(
  GPS_data = GPS_data,
  interval = interval,
  column_datetime = column_datetime,
  column_lat = column_lat,
  column_lon = column_lon
)
}
\arguments{
\item{GPS_data}{you GPS data}

\item{interval}{the interval in seconds}

\item{column_datetime}{a column with date and time, and in a date time format}

\item{column_lat}{the name of your column with the 'Latitude' data}

\item{column_lon}{the name of your column with the 'Longitude' data}
}
\value{
returns a interpolated data frame using the interval provided
}
\description{
A function to interpolate your data at an specific interval
}
\examples{
GPS01_trip5 <-GPS_preparado[GPS_preparado$ID == 'GPS01' & GPS_preparado$trip_number=='trip_5',]
GPS01_trip1_interpolated<-interpolate(GPS_data = GPS01_trip5,interval='10 sec',
column_datetime = 'dia_hora',column_lat = 'Latitude',column_lon = 'Longitude')
}
