% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_trips.R
\name{interpolate_trips}
\alias{interpolate_trips}
\title{interpolate_trips, interpolate your tracking data to a specific interval}
\usage{
interpolate_trips(
  GPS_data = GPS_data,
  interval = interval,
  column_date = column_date,
  column_time = column_time,
  datetime_format = datetime_format,
  column_lat = column_lat,
  column_lon = column_lon,
  column_trip = column_trip
)
}
\arguments{
\item{GPS_data}{Is the data you want to interpolate}

\item{interval}{Is the interval you want to use, in seconds, for example 900 s is 15 minutes}

\item{column_date}{Is the name of the column where you have the Date, if you dont have date separately, you can always create a new column only with the date}

\item{column_time}{Is the name of the column where you have the Time, if you dont have time separately in a column, you can always create a new column only with the time}

\item{datetime_format}{Is the format you use for your date and time, for example "\%d/\%m/\%Y \%H:\%M:\%S", "\%d-\%m-\%Y \%H:\%M:\%S" or "\%Y-\%m-\%d \%H:\%M:\%S"}

\item{column_lat}{Is the name of the column where you have the Latitude}

\item{column_lon}{Is the name of the column where you have the Longitude}

\item{column_trip}{Is the name of the column where you have the Trip number, if you dont have number of trip please check the function count_trips}
}
\value{
Gives back a data frame with the interpolated data, only four columns will be returned: dt, Longitude, Latitude, and trip_number
}
\description{
interpolate_trips, interpolate your tracking data to a specific interval
}
\examples{
GPS_interpolated<-interpolate_trips(GPS_data=GPS01_trips,interval='900 sec',
column_date='DateGMT',column_time='TimeGMT',column_trip='trip_number',
column_lat='Latitude',column_lon='Longitude',datetime_format="\%d/\%m/\%Y \%H:\%M:\%S")
}
