% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localizar_nidos.R
\name{localizar_nidos}
\alias{localizar_nidos}
\title{Obtiene informacion del nido, usando el primer punto de registro}
\usage{
localizar_nidos(
  GPS_data = GPS_data,
  lon_col = lon_col,
  lat_col = lat_col,
  ID_col = ID_col
)
}
\arguments{
\item{GPS_data}{Un data frame con las columnas Latitude, Longitude y IDs.}

\item{lon_col}{El nombre de la columna donde esta la Longitude. Debe escribirse entre comillas.}

\item{lat_col}{El nombre de la columna donde esta la Latitude. Debe escribirse entre comillas.}

\item{ID_col}{El nombre de la columna donde estan los IDs. Debe escribirse entre comillas.}
}
\value{
Un data frame con la localizacion de los nidos
}
\description{
Obtiene informacion del nido, usando el primer punto de registro
}
\examples{
Nidos_df<-localizar_nidos(GPS_data=GPS_raw,lon_col="Longitude",lat_col="Latitude",ID_col="IDs")
}
