% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_nest.R
\name{locate_nest}
\alias{locate_nest}
\title{Use first location to identify nest location}
\usage{
locate_nest(
  GPS_data = GPS_data,
  column_lon = column_lon,
  column_lat = column_lat
)
}
\arguments{
\item{GPS_data}{A data frame with GPS data}

\item{column_lon}{The name of the column where the longitude data is in your data frame.}

\item{column_lat}{The name of the column where the latitude data is your data frame.}
}
\value{
The nest location
}
\description{
Use first location to identify nest location
}
\examples{
GPS01_nest<-locate_nest(GPS_data = GPS01,column_lat = 'Longitude',column_lon = 'Latitude')
}
