% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Notas.R
\docType{data}
\name{Notas}
\alias{Notas}
\title{Datos de libreta de bobo/piquero/kena Sula dactylatra en Rapa Nui}
\format{
Un data frame en formato rda, contiene datos de diez individuos.
Las columnas son IDs, Hora_inicio, Hora_final
}
\usage{
data(Notas)
}
\description{
Datos de libreta hechos durante muestreo de Sula dactylatra en Rapa Nui
Incluye 10 observaciones de 3 variables
Incluye datos de 10 individuos
}
\examples{
data(Notas)
}
\references{
Lerma et al. (2020)
}
\keyword{datasets}
