% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparar_varios.R
\name{preparar_varios}
\alias{preparar_varios}
\title{Identifica viajes, calcula distancia de la colonia de los puntos y distancia entre puntos}
\usage{
preparar_varios(
  GPS_data = GPS_data,
  lon_col = lon_col,
  lat_col = lat_col,
  ID_col = ID_col,
  distancia_km = distancia_km,
  sistema_ref = sistema_ref
)
}
\arguments{
\item{GPS_data}{un data frame con columnas longitud y latitud en decimales}

\item{lon_col}{el nombre de la columna con la longitud, es necesario escribirlo con comillas}

\item{lat_col}{el nombre de la columna con la latitud, escribir con comillas}

\item{ID_col}{el nombre de la columna con los IDs}

\item{distancia_km}{un numero correspondiente a la distancia de la colonia}

\item{sistema_ref}{sistema de referencia, por ejemplo "+init=epsg:4326"}
}
\value{
el mismo data frame con la columna trip_number, num_seq, max_dist_km, y pointsdist_km
}
\description{
Identifica viajes, calcula distancia de la colonia de los puntos y distancia entre puntos
}
\examples{
GPS_final<-preparar_varios(GPS_data=GPS_raw,
lon_col="Longitude",
lat_col="Latitude",
ID_col="IDs",
distancia_km=1,
sistema_ref="+init=epsg:4326")
}
