% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recortar_periodo.R
\name{recortar_periodo}
\alias{recortar_periodo}
\title{Recorta periodos dentro de los datos}
\usage{
recortar_periodo(
  GPS_data = GPS_data,
  dia_col = dia_col,
  hora_col = hora_col,
  formato = formato,
  inicio = inicio,
  final = final
)
}
\arguments{
\item{GPS_data}{data frame con columna hora y dia}

\item{dia_col}{nombre de la columna donde esta el dia}

\item{hora_col}{nombre de la columna donde esta la hora}

\item{formato}{formato de hora y dia}

\item{inicio}{inicio del periodo que se quiere recortar}

\item{final}{final del periodo que se quiere recortar}
}
\value{
el mismo data frame solo con el periodo que interesa
}
\description{
Recorta periodos dentro de los datos
}
\examples{
GPS_recortado<-recortar_periodo(GPS_data=GPS_01,
inicio='02/11/2017 18:10:00',
final='05/11/2017 14:10:00',
dia_col='DateGMT',
hora_col='TimeGMT',
formato="\%d/\%m/\%Y \%H:\%M:\%S")
}
