% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recortar_por_ID.R
\name{recortar_por_ID}
\alias{recortar_por_ID}
\title{Recorta periodos usando informacion de IDs}
\usage{
recortar_por_ID(GPS_data = GPS_data, Notas = Notas, formato = formato)
}
\arguments{
\item{GPS_data}{un data frame con las columnas 'IDs','DateGMT' y 'TimeGMT' con estos nombres}

\item{Notas}{un data frame con las columnas 'Hora_inicio' y 'Hora_final' con estos nombres}

\item{formato}{el formato en el que estan escritas las fechas, debe ser el mismo para ambos data frames}
}
\value{
el data frame GPS_data pero recortado, es decir con menos observaciones
}
\description{
Recorta periodos usando informacion de IDs
}
\examples{
GPS_recortados<-recortar_por_ID(GPS_data=GPS_raw,Notas=Notas,formato="\%d/\%m/\%Y \%H:\%M:\%S")
}
