% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajustar_hora.R
\name{ajustar_hora}
\alias{ajustar_hora}
\title{Corrige la hora en tu base de datos}
\usage{
ajustar_hora(datos, columna, diferencia_horaria)
}
\arguments{
\item{datos}{Un data frame que contenga una columna que tenga datos en POSIXct}

\item{columna}{Columna con fecha y hora, para convertir hay que pegar la columa dia y hora y usar la funcion as.POSIXct}

\item{diferencia_horaria}{El numero de horas de diferencia}
}
\value{
Regresa un nuevo data frame con el nombre 'horas_corregidas'
}
\description{
Corrige la hora en tu base de datos
}
\examples{
ajustar_hora(datos=GPSdata2, columna=diahora,diferencia_horaria=5)
}
