% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajustar_hora.R
\name{ajustar_hora}
\alias{ajustar_hora}
\title{Corrige el tiempo de acuerdo a la zona horaria.}
\usage{
ajustar_hora(
  GPS_data = GPS_data,
  dif_hor = dif_hor,
  dia_col = dia_col,
  hora_col = hora_col,
  formato = formato
)
}
\arguments{
\item{GPS_data}{tu data frame con columna de hora y tiempo}

\item{dif_hor}{un numero correspondiente a la diferencia horaria}

\item{dia_col}{el nombre de la columna donde tienes la fecha}

\item{hora_col}{el nombre de la columna donde tienes el tiempo}

\item{formato}{el formato en el que esta tu fecha y hora}
}
\value{
un data frame con dos columnas adicionales dia_hora y hora_corregida
}
\description{
Corrige el tiempo de acuerdo a la zona horaria.
}
\examples{
t_formato="\%d/\%m/\%Y \%H:\%M:\%S"
GPS_gmt<-ajustar_hora(GPS_data = GPS_raw, dif_hor = 5,
dia_col = 'DateGMT', hora_col = 'TimeGMT',formato=t_formato)
}
