% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_duracion.R
\name{calcular_duracion}
\alias{calcular_duracion}
\title{Calcula la duracion de los viajes de alimentacion}
\usage{
calcular_duracion(
  GPS_data = GPS_data,
  col_diahora = col_diahora,
  formato = formato,
  unidades = unidades
)
}
\arguments{
\item{GPS_data}{un data frame con una columna que incluya dia y hora, y otra columna que incluya el numero del viaje}

\item{col_diahora}{la columna en formato POSTIXct con informacion del dia y hora}

\item{formato}{el formato en el que esta la hora y dia en la columna POSTIXct con informacion del dia y hora}

\item{unidades}{elegir "hours", "minutes", "seconds".}
}
\value{
regresa un nuevo data frame con la informacion del viaje, cuando inicio, cuando termino y la duracion
}
\description{
Calcula la duracion de los viajes de alimentacion
}
\examples{
formato<-"\%Y-\%m-\%d \%H:\%M:\%S"
duracion<-calcular_duracion(GPS_data=GPS_edited,col_diahora="tStamp",
formato=formato,unidades="hours")
}
