% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolar_viajes.R
\name{interpolar_viajes}
\alias{interpolar_viajes}
\title{Interpola los viajes de acuerdo a intervalo}
\usage{
interpolar_viajes(
  GPS_preparado = GPS_preparado,
  intervalo = intervalo,
  col_diahora = col_diahora,
  col_ID = col_ID,
  separador = separador
)
}
\arguments{
\item{GPS_preparado}{es el data frame con columnas "Latitude", "Longitude" y "tStamp"}

\item{intervalo}{es el intervalo al que quieres convertir tu track, por ejemplo "900 secs"}

\item{col_diahora}{es la columna con la informacion de dia y hora, en formato POSIXct, por ejemplo "tStamp"}

\item{col_ID}{es la columna con la informacion de los invididuos, por ejemplo "ID"}

\item{separador}{es la columna con la informacion de los viajes, por ejemplo 'trip_number'}
}
\value{
el data frame con datos a cada intervalo de tiempo
}
\description{
Interpola los viajes de acuerdo a intervalo
}
\examples{
GPS_interpolated<-interpolar_viajes(GPS_preparado=GPS_preparado, 
intervalo="900 sec", 
col_diahora="dia_hora", 
separador='trip_number',
col_ID='IDs')

}
